+++ Sources/ConsoleDockCore/CDKConfiguration.m
@@ -7,6 +7,7 @@
     self = [super init];
     if (self) {
         _maximumEntries = 2000;
+        _maximumMessageLength = 8192;
         _captureStandardOutput = YES;
         _captureStandardError = YES;
         _showsFloatingButton = YES;
@@ -24,10 +25,12 @@
 {
     CDKConfiguration *copy = [[[self class] allocWithZone:zone] init];
     copy.maximumEntries = self.maximumEntries;
+    copy.maximumMessageLength = self.maximumMessageLength;
     copy.captureStandardOutput = self.captureStandardOutput;
     copy.captureStandardError = self.captureStandardError;
     copy.showsFloatingButton = self.showsFloatingButton;
     copy.allowsReleaseBuilds = self.allowsReleaseBuilds;
+    copy.redactionBlock = self.redactionBlock;
     return copy;
 }