+++ Sources/ConsoleDockCore/include/ConsoleDockCore.h
@@ -10,6 +10,12 @@ typedef NS_ENUM(NSInteger, CDKLogLevel) {
     CDKLogLevelFault
 };
 
+typedef NS_ENUM(NSInteger, CDKLogSource) {
+    CDKLogSourceNative = 0,
+    CDKLogSourceStdout,
+    CDKLogSourceStderr
+};
+
 typedef NS_ENUM(NSInteger, CDKStartResult) {
     CDKStartResultStarted = 0,
     CDKStartResultAlreadyRunning,
@@ -17,17 +23,37 @@ typedef NS_ENUM(NSInteger, CDKStartResult) {
     CDKStartResultFailed
 };
 
+typedef NSString * _Nonnull (^CDKRedactionBlock)(NSString *message);
+
 FOUNDATION_EXPORT NSErrorDomain const CDKConsoleDockErrorDomain;
 
 @interface CDKConfiguration : NSObject <NSCopying>
 
 @property (nonatomic) NSUInteger maximumEntries;
+@property (nonatomic) NSUInteger maximumMessageLength;
 @property (nonatomic) BOOL captureStandardOutput;
 @property (nonatomic) BOOL captureStandardError;
 @property (nonatomic) BOOL showsFloatingButton;
 @property (nonatomic) BOOL allowsReleaseBuilds;
+@property (nonatomic, copy, nullable) CDKRedactionBlock redactionBlock;
 
 + (instancetype)defaultConfiguration;
+
+@end
+
+@interface CDKLogEntry : NSObject <NSCopying>
+
+@property (nonatomic, copy, readonly) NSDate *timestamp;
+@property (nonatomic, readonly) CDKLogLevel level;
+@property (nonatomic, readonly) CDKLogSource source;
+@property (nonatomic, copy, readonly) NSString *message;
+
+- (instancetype)init NS_UNAVAILABLE;
++ (instancetype)new NS_UNAVAILABLE;
+- (instancetype)initWithTimestamp:(NSDate *)timestamp
+                            level:(CDKLogLevel)level
+                           source:(CDKLogSource)source
+                          message:(NSString *)message NS_DESIGNATED_INITIALIZER;
 
 @end
 
@@ -38,6 +64,9 @@ FOUNDATION_EXPORT NSErrorDomain const CDKConsoleDockErrorDomain;
                                    error:(NSError * _Nullable * _Nullable)error;
 + (void)stop;
 + (BOOL)isRunning;
+
++ (NSArray<CDKLogEntry *> *)entries;
++ (void)clearEntries;
 
 + (void)logWithLevel:(CDKLogLevel)level message:(NSString *)message;
 + (void)debug:(NSString *)message;